Reorganize code
* Create baseui module to reduce code duplication
* Edit Makefile to run the webapps
