+++ components/jupyter-web-app/default/kubeflow/jupyterui/__init__.py
@@ -2,4 +2,4 @@ from flask import Flask, url_for, jsonify
 
 app = Flask(__name__)
 
-from kubeflow.jupyter import routes
+from kubeflow.jupyterui import routes
