+++ components/jupyter-web-app/default/kubeflow/jupyterui/static/css/notebooks.css
