+++ components/jupyter-web-app/default/kubeflow/jupyterui/static/js/add_notebook.js
