+++ components/jupyter-web-app/default/kubeflow/jupyterui/static/js/notebooks.js
