+++ components/jupyter-web-app/default/kubeflow/jupyterui/static/kubeflow.png
