+++ components/jupyter-web-app/default/kubeflow/jupyterui/templates/add_notebook.html
