+++ components/jupyter-web-app/default/run.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from kubeflow.jupyter import app
+from kubeflow.jupyterui import app
 
 if __name__ == '__main__':
   app.run(debug=True, host="0.0.0.0")
